fixed some issues here and there
