--- src/types/FacetType.ts
@@ -1,9 +1,0 @@
-export interface FacetResultType {
-	catalog: string[];
-	collection: string[];
-	genre: string[];
-	resource_description: string[];
-	resource_description_general: string[];
-	creator_full_name: string[];
-	subject_full_name: string[];
-}
